@@ -14,7 +14,7 @@
     "app": "0.0.0",
     "google-auth-library": "^9.4.1",
     "google-spreadsheet": "^4.1.1",
-    "next": "13.4.4",
+    "next": "^14.0.4",
     "raf": "^3.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
