@@ -1,12 +1,12 @@
 import React from 'react'
-import { Text, View } from 'tamagui'
+import { Text, View } from '@my/ui'
 
-type TableBodyProps = {
-  past: boolean
+export type TableProps = {
+  past?: boolean
   children: React.ReactNode
 }
 
-export const TableBody: React.FC<TableBodyProps> = ({ children, past }) => {
+export const TableBody: React.FC<TableProps> = ({ children, past }) => {
   return (
     <View flex={1} flexBasis={0}>
       <Text fontStyle={'normal'} color={past ? '$gray12Dark' : '$grey2Dark'}>
